Break out tests into separate file
