+++ contracts/distribution/dao-rewards-distributor/src/lib.rs
@@ -5,4 +5,7 @@ mod error;
 pub mod msg;
 pub mod state;
 
+#[cfg(test)]
+mod tests;
+
 pub use crate::error::ContractError;
